+++ JewelryEC-Backend/JewelryEC-Backend/Models/Carts/Dto/GetCartResponseDto.cs
@@ -10,7 +10,7 @@ namespace JewelryEC_Backend.Models.Carts.Dto
         public int IsPayed { get; set; } = 0;
         public string? UserId { get; set; }
         public List<GetCartItemResponseDto> ? Items { get; set; }
-        public float? TotalPrice { get; set; }
+        public decimal? TotalPrice { get; set; }
         public GetCartResponseDto()
         {
             // Khởi tạo TotalPrice là 0
